Fixed issues with execute/computeResults methods  in geoserver commands

+++ extensions/cli/geoserver/src/main/java/mil/nga/giat/geowave/cli/geoserver/GeoServerRemoveFeatureLayerCommand.java
@@ -38,11 +38,6 @@ public class GeoServerRemoveFeatureLayerCommand extends
 	public void execute(
 			final OperationParams params )
 			throws Exception {
-		if (parameters.size() != 1) {
-			throw new ParameterException(
-					"Requires argument: <layer name>");
-		}
-
 		JCommander.getConsole().println(
 				computeResults(params));
 	}
@@ -51,6 +46,11 @@ public class GeoServerRemoveFeatureLayerCommand extends
 	public String computeResults(
 			final OperationParams params )
 			throws Exception {
+		if (parameters.size() != 1) {
+			throw new ParameterException(
+					"Requires argument: <layer name>");
+		}
+		
 		layerName = parameters.get(0);
 
 		final Response deleteLayerResponse = geoserverClient.deleteFeatureLayer(layerName);
